feat: Initial WebSocket implementation
Initial websocket implementation, based on channel ids.
Implemented a way to keep track currently connected users onto a given websocket session, which allows for functionality such as kicking users from a channel, without the client being able to access the messages if they do not unsubscribe from the channel manually. (There might be a better way of doing this?)

TODO: Verify if the user belongs to a given channel whenever subscribing or sending a message.