fix: Explicit mapping for endpoints that need to be authenticated
Previous implementation of the securityFilterChain made it so every URL needed to have users authenticated, except for specific urls specified. With this change, every url can have unauthenticated users with the exception of specified ones. This fixed the issue where every endpoint returns 403, even if the endpoint actually does not exist, which helps in debugging.

You can now send events to users that are currently subscribed to the event topic via the UserEventHandler service.